refactor: move projector scoring math to core/scoring
