+++ src/quodeq/core/scoring/projector_scoring.py
